fix: Support different query parameter names in search
This fixes an issue that easily reproducible with KOReader, but might
affect other OPDS clients too.

Until now, we expected the client to define a `searchTerms` query
parameter. However, according to the OPDS spec at

https://specs.opds.io/opds-1.2#3-search

the query parameter should actually be called `q` instead. Even more
confusing is, that KOReader uses `query`.

For backwards compatibility reasons we keep `searchTerms` and add the
other two options as well. All parameters are now optional, so we need
to check which one we got.

I decided to return a 400 Bad Request in case none of the query
parameters was given. We can determine this in case all of them are the
default value of `null`.

Returning a Bad Request error ensures that we will detect issues with
other OPDS clients in the future, in case there are implementations out
there that use _yet another_ parameter name for the search terms.

If we instead processed the request as empty by default, a client that
uses a different query parameter name would just have the entire catalog
returned. Users would be confused why their filters did not work and the
underlying issue might be hard to track down.

This fixes #40.